generalize-optimization-algorithm
generalize genetic-algorithm into optimizer,
extracting all optimization-related responsibilities to Optimizer,
and leaving OptimizationAlgorithm with only performing optimization-step
