+++ src/main/java/examples/bit/vector/BitVectorOptimization.java
@@ -1,11 +1,12 @@
 package examples.bit.vector;
 
-import optimization.genetic.GeneticAlgorithm;
-import optimization.genetic.GeneticAlgorithmBuilder;
-import optimization.genetic.listener.StatisticsCollector;
-import optimization.genetic.operator.mutation.FixedRateMutation;
-import optimization.genetic.selection.Selector;
-import optimization.genetic.selection.roulette.wheel.RouletteWheelSelector;
+import optimization.OptimizationAlgorithm;
+import optimization.OptimizerBuilder;
+import optimization.algorithm.genetic.GeneticAlgorithmBuilder;
+import optimization.algorithm.genetic.operator.mutation.FixedRateMutation;
+import optimization.algorithm.genetic.selection.roulette.wheel.RouletteWheelSelector;
+import optimization.evaluation.OptimizationStepEvaluation;
+import optimization.step.listener.StatisticsLogger;
 import termination.LimitedEpochsTerminationCriterion;
 
 import java.util.Collections;
@@ -14,18 +15,24 @@ import java.util.Random;
 
 public class BitVectorOptimization {
 	public static void main(String[] args) {
-		Selector<String> result = new GeneticAlgorithmBuilder<>(List.of("000", "001", "100", "110", "011"))
+		OptimizationStepEvaluation<String> result = new OptimizerBuilder<>(List.of("000", "001", "100", "110", "011"))
 				.maximize()
-				.fitnessFunction(str -> (float) str.chars().filter(i -> i == '1').count())
+				.evaluationFunction(str -> (float) str.chars().filter(i -> i == '1').count())
+				.listeners(Collections.singletonList(new StatisticsLogger<>()))
+				.terminationCriterion(new LimitedEpochsTerminationCriterion<>(5))
+				.optimizationAlgorithm(createOptimizationAlgorithm())
+				.build().optimize();
+
+		result.getBestSolutions(3).forEach(System.out::println);
+	}
+
+	private static OptimizationAlgorithm<String> createOptimizationAlgorithm() {
+		return new GeneticAlgorithmBuilder<String>()
 				.selectorFactory(RouletteWheelSelector::new)
 				.crossoverFunction(selector -> selector.select().charAt(0) + selector.select().substring(1))
-				.elitismSize(1)
 				.geneticOperator(createStringMutation())
-				.terminationCriterion(new LimitedEpochsTerminationCriterion<>(5, GeneticAlgorithm::getGeneration))
-				.generationListeners(Collections.singletonList(new StatisticsCollector<>()))
-				.build().run();
-
-		result.getElitism(3).forEach(System.out::println);
+				.elitismSize(1)
+				.build();
 	}
 
 	private static FixedRateMutation<String> createStringMutation() {
