+++ src/main/java/optimization/OptimizationAlgorithm.java
@@ -1,0 +1,9 @@
+package optimization;
+
+import optimization.evaluation.OptimizationStepEvaluation;
+
+import java.util.Collection;
+
+public interface OptimizationAlgorithm<T> {
+	Collection<T> optimize(OptimizationStepEvaluation<T> stepEvaluation);
+}
