+++ src/main/java/optimization/algorithm/genetic/CrossoverFunction.java
@@ -1,0 +1,8 @@
+package optimization.algorithm.genetic;
+
+import optimization.algorithm.genetic.selection.Selector;
+
+@FunctionalInterface
+public interface CrossoverFunction<T> {
+	T createOffspringBy(Selector<T> selector);
+}
