+++ src/main/java/optimization/algorithm/genetic/operator/GeneticOperator.java
@@ -1,0 +1,6 @@
+package optimization.algorithm.genetic.operator;
+
+@FunctionalInterface
+public interface GeneticOperator<T> {
+	T apply(T individual);
+}
