+++ src/main/java/optimization/algorithm/genetic/selection/Selector.java
@@ -1,0 +1,5 @@
+package optimization.algorithm.genetic.selection;
+
+public interface Selector<T> {
+	T select();
+}
