+++ src/main/java/optimization/algorithm/genetic/selection/SelectorFactory.java
@@ -1,0 +1,8 @@
+package optimization.algorithm.genetic.selection;
+
+import java.util.Map;
+
+@FunctionalInterface
+public interface SelectorFactory<T> {
+	Selector<T> create(Map<T, Float> fitnessScores, boolean isMinimization);
+}
