+++ src/main/java/optimization/evaluation/EvaluationFunction.java
@@ -1,0 +1,6 @@
+package optimization.evaluation;
+
+@FunctionalInterface
+public interface EvaluationFunction<T> {
+	Float evaluate(T object);
+}
