--- src/main/java/optimization/genetic/CrossoverFunction.java
@@ -1,8 +1,0 @@
-package optimization.genetic;
-
-import optimization.genetic.selection.Selector;
-
-@FunctionalInterface
-public interface CrossoverFunction<T> {
-	T createOffspringBy(Selector<T> selector);
-}
