--- src/main/java/optimization/genetic/FitnessFunction.java
@@ -1,6 +1,0 @@
-package optimization.genetic;
-
-@FunctionalInterface
-public interface FitnessFunction<T> {
-	Float evaluate(T individual);
-}
