--- src/main/java/optimization/genetic/operator/GeneticOperator.java
@@ -1,6 +1,0 @@
-package optimization.genetic.operator;
-
-@FunctionalInterface
-public interface GeneticOperator<T> {
-	T apply(T individual);
-}
