--- src/main/java/optimization/genetic/selection/SelectorFactory.java
@@ -1,8 +1,0 @@
-package optimization.genetic.selection;
-
-import java.util.Map;
-
-@FunctionalInterface
-public interface SelectorFactory<T> {
-	Selector<T> create(Map<T, Float> fitnessScores, boolean isMinimization);
-}
