+++ src/main/java/optimization/step/listener/OptimizationStepListener.java
@@ -1,0 +1,9 @@
+package optimization.step.listener;
+
+import optimization.evaluation.OptimizationStepEvaluation;
+
+public interface OptimizationStepListener<T> {
+	void notifyOptimizationStepStarted();
+
+	void notifyOptimizationStepEnded(OptimizationStepEvaluation<T> stepEvaluation);
+}
