+++ src/main/java/termination/LimitedEpochsTerminationCriterion.java
@@ -1,15 +1,11 @@
 package termination;
 
-import java.util.function.Function;
-
 public class LimitedEpochsTerminationCriterion<T> implements TerminationCriterion<T> {
 	private int numEpochs;
-	private Function<T, Integer> epochSupplier;
 
-	public LimitedEpochsTerminationCriterion(int numEpochs, Function<T, Integer> epochSupplier) {
+	public LimitedEpochsTerminationCriterion(int numEpochs) {
 		validateNumGenerations(numEpochs);
 		this.numEpochs = numEpochs;
-		this.epochSupplier = epochSupplier;
 	}
 
 	private void validateNumGenerations(int numGenerations) {
@@ -20,6 +16,6 @@ public class LimitedEpochsTerminationCriterion<T> implements TerminationCriterio
 
 	@Override
 	public boolean isDone(T object) {
-		return epochSupplier.apply(object) >= numEpochs;
+		return --numEpochs <= 0;
 	}
 }
